==> 3_uncertainty_analyses/01_calculate_GUMBOOT_sampling_uncertainty.ipynb <==
# Calculate GUMBOOT sampling uncertainty
### Jackknife and bootstrap

# Load Package
# install.packages("gumboot")
library('gumboot')

## Adjusted model code in Gumboot package

## Adjusted code in package Bootjack function trace('bootjack', edit=TRUE):
# qSimValid = qSimValid[!is.na(qSimValid)]
# qObsValid = qObsValid[!is.na(qObsValid)]


# rankSim <- order(order(qSimValid)) -1
# rankObs <- order(order(qObsValid)) -1


# r_num <- sum((rankObs - mean(rankObs)) * (rankSim - mean(rankSim)))
# r_den <- sqrt(sum((rankObs - mean(rankObs))^2) * sum((rankSim - mean(rankSim))^2))
# r <- r_num / r_den

# alpha <- 1 - 0.5 * sum(abs(fdcSim - fdcObs))
# beta <-  mean(qSimValid)/ mean(qObsValid)

# kge_np <- 1 - sqrt((r - 1) ^2 + (alpha - 1)^2 + (beta - 1)^2)
# kge_np

# wflow_sbm Calibrated GUMBOOT

# Evaluation result files
files <- Sys.glob(sprintf("/gpfs/work1/0/wtrcycle/users/jaerts/camels_uk/results/wflow_sbm/evaluation_period_calibrated/gumboot/*_evaluation_simulations.csv"))

## KGE-NP

# Create empty dataframe
df_total = data.frame()

# Loop files
for (file in files){
    
    # Load data
    df <- read.csv(file)
    df['date'] <- as.Date(df$date)
    
    # Select info
    basin_id = strsplit(file,'/')[[1]][13]
    basin_id = strsplit(basin_id,'_')[[1]][1]
    
    
    
    # Calculate bootjack
    data <- bootjack(flows = df,
             GOF_stat = 'KGE',
             nSample = 500,
             waterYearMonth=13,
             seed=1000,
             minDays = 100,
             minYears = 5)
    
    df_data <- data.frame(data)
    df_data$basin_id <- basin_id
    df_total <- rbind(df_total,df_data)
}

write.csv(df_total,"/gpfs/work1/0/wtrcycle/users/jaerts/camels_uk/results/wflow_sbm/wflow_calibrated_kge_np_gumboot.csv", row.names = FALSE)

## NSE

# Create empty dataframe
df_total = data.frame()

# Loop files
for (file in files){
    
    # Load data
    df <- read.csv(file)
    df['date'] <- as.Date(df$date)
    
    # Select info
    basin_id = strsplit(file,'/')[[1]][13]
    basin_id = strsplit(basin_id,'_')[[1]][1]
    
    
    
    # Calculate bootjack
    data <- bootjack(flows = df,
             GOF_stat = 'NSE',
             nSample = 500,
             waterYearMonth=13,
             seed=1000,
             minDays = 100,
             minYears = 5)
    
    df_data <- data.frame(data)
    df_data$basin_id <- basin_id
    df_total <- rbind(df_total,df_data)
}

write.csv(df_total,"/gpfs/work1/0/wtrcycle/users/jaerts/camels_uk/results/wflow_sbm/wflow_calibrated_nse_gumboot.csv", row.names = FALSE)

# wflow_sbm Uncalibrated GUMBOOT

# Evaluation result files
files <- Sys.glob(sprintf("/gpfs/work1/0/wtrcycle/users/jaerts/camels_uk/results/wflow_sbm/evaluation_period_uncalibrated/gumboot/*_evaluation_simulations.csv"))

## KGE-NP

# Create empty dataframe
df_total = data.frame()

# Loop files
for (file in files){

    # Load data
    df <- read.csv(file)
    df['date'] <- as.Date(df$date)
    
    # Select info
    basin_id = strsplit(file,'/')[[1]][13]
    basin_id = strsplit(basin_id,'_')[[1]][1]
    
    
    
    # Calculate bootjack
    data <- bootjack(flows = df,
             GOF_stat = 'KGE',
             nSample = 500,
             waterYearMonth=13,
             seed=1000,
             minDays = 100,
             minYears = 5)
    
    df_data <- data.frame(data)
    df_data$basin_id <- basin_id
    df_total <- rbind(df_total,df_data)
}

write.csv(df_total,"/gpfs/work1/0/wtrcycle/users/jaerts/camels_uk/results/wflow_sbm/wflow_uncalibrated_kge_np_gumboot.csv", row.names = FALSE)

## NSE

# Create empty dataframe
df_total = data.frame()

# Loop files
for (file in files){
    
    # Load data
    df <- read.csv(file)
    df['date'] <- as.Date(df$date)
    
    # Select info
    basin_id = strsplit(file,'/')[[1]][13]
    basin_id = strsplit(basin_id,'_')[[1]][1]
    
    
    
    # Calculate bootjack
    data <- bootjack(flows = df,
             GOF_stat = 'NSE',
             nSample = 500,
             waterYearMonth=13,
             seed=1000,
             minDays = 100,
             minYears = 5)
    
    df_data <- data.frame(data)
    df_data$basin_id <- basin_id
    df_total <- rbind(df_total,df_data)
}

write.csv(df_total,"/gpfs/work1/0/wtrcycle/users/jaerts/camels_uk/results/wflow_sbm/wflow_uncalibrated_nse_gumboot.csv", row.names = FALSE)

# PCR-GLOBWB

# Evaluation result files
files <- Sys.glob(sprintf("/gpfs/work1/0/wtrcycle/users/jaerts/camels_uk/results/pcr-globwb/evaluation_period/gumboot/*_gumboot_pcr-globwb_evaluation_simulations.csv"))

## KGE-NP

# Create empty dataframe
df_total = data.frame()

# Loop files
for (file in files){

    # Load data
    df <- read.csv(file)
    df['date'] <- as.Date(df$date)
    
    # Select info
    basin_id = strsplit(file,'/')[[1]][13]
    basin_id = strsplit(basin_id,'_')[[1]][1]
    
    
    
    # Calculate bootjack
    data <- bootjack(flows = df,
             GOF_stat = 'KGE',
             nSample = 500,
             waterYearMonth=13,
             seed=1000,
             minDays = 100,
             minYears = 5)
    
    df_data <- data.frame(data)
    df_data$basin_id <- basin_id
    df_total <- rbind(df_total,df_data)
}

write.csv(df_total,"/gpfs/work1/0/wtrcycle/users/jaerts/camels_uk/results/pcr-globwb/pcr-globwb_kge_np_gumboot.csv", row.names = FALSE)

## NSE

# Create empty dataframe
df_total = data.frame()

# Loop files
for (file in files){
    
    # Load data
    df <- read.csv(file)
    df['date'] <- as.Date(df$date)
    
    # Select info
    basin_id = strsplit(file,'/')[[1]][13]
    basin_id = strsplit(basin_id,'_')[[1]][1]
    
    
    
    # Calculate bootjack
    data <- bootjack(flows = df,
             GOF_stat = 'NSE',
             nSample = 500,
             waterYearMonth=13,
             seed=1000,
             minDays = 100,
             minYears = 5)
    
    df_data <- data.frame(data)
    df_data$basin_id <- basin_id
    df_total <- rbind(df_total,df_data)
}

write.csv(df_total,"/gpfs/work1/0/wtrcycle/users/jaerts/camels_uk/results/pcr-globwb/pcr-globwb_nse_gumboot.csv", row.names = FALSE)
